ai/requirements.txt
streamlit
torch
torchvision
pillow
pandas
matplotlib
openai